Upgrade from argparse to click
- Upgrade from argparse to click
- Added unit test
- Added config file location command
