+++ MANIFEST.in
@@ -3,6 +3,7 @@ include CONTRIBUTING.rst
 include HISTORY.rst
 include LICENSE
 include README.rst
+include picturebot/config.json
 
 recursive-include tests *
 recursive-exclude * __pycache__
