+++ picturebot/__init__.py
@@ -4,4 +4,4 @@
 
 __author__ = """Tomek Joostens"""
 __email__ = 'joostenstomek@gmail.com'
-__version__ = '0.0.1'
+__version__ = '0.0.2'
