+++ requirements_dev.txt
@@ -7,5 +7,5 @@ tox==3.5.2
 coverage==4.5.1
 Sphinx==1.8.1
 twine==1.12.1
-
+generalutils
