+++ tests/test_picturebot.py
@@ -6,11 +6,29 @@
 import os
 import random
 import unittest
-
+import picturebot.helper as helper
+import generalutils.guard as grd
 class TestGeneralUtils(unittest.TestCase):
 
-    def test_FullPath(self):
-        pass
+    def test_FullFilePathExists(self):
+        '''Test the full file path'''
+
+        # Obtain the expected full path
+        pathExpected = os.path.join(os.path.dirname(helper.__file__), 'qq', 'ww', 'ee')
+        # Obtain the returned path from the FullFilePathFunction
+        pathTest = helper.FullFilePath('qq', 'ww', 'ee')
+
+        self.assertEqual(pathExpected, pathTest, f'{pathTest} is not equal to {pathExpected}')
+
+    def test_FullFilePathDoesNotExists(self):
+        '''Negative test to check whether a certain full file path doesn't exists'''
+
+        # Obtain the expected full path
+        pathExpected = os.path.join(os.getcwd(), 'qq', 'ww', 'ee')
+        # Obtain the returned path from the FullFilePathFunction
+        pathTest = helper.FullFilePath('qq', 'ww', 'ee')
+
+        self.assertNotEqual(pathExpected, pathTest, f'{pathTest} are equal {pathExpected}')
 
 if __name__ == '__main__':
 	unittest.main()
